pyproject.toml
[project]
name = "finrag"
version = "0.1.0"
description = "Financial Document RAG + MCP Server Agent using Google ADK and FastMCP"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.5.0",
    "google-genai>=1.0.0",
    "fastmcp>=2.0.0",
    "chromadb>=0.5.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.setuptools.packages.find]
include = ["rag*", "config*", "agents*", "mcp_server*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
